+++ unit_tests/aub_tests/gen9/execution_model/enqueue_parent_kernel_tests_gen9.cpp
@@ -74,10 +74,10 @@ GEN9TEST_F(AUBParentKernelFixture, EnqueueParentKernel) {
             CL_FILTER_LINEAR,
             retVal);
 
-        size_t argScalar = 2;
+        uint64_t argScalar = 2;
         pKernel->setArg(
             3,
-            sizeof(size_t),
+            sizeof(uint64_t),
             &argScalar);
 
         pKernel->setArg(
